--- app/src/main/java/com/xmartlabs/taskloans/helper/OnBoardingValidator.kt
@@ -1,16 +1,0 @@
-package com.xmartlabs.taskloans.helper
-
-object OnBoardingValidator {
-    private const val MIN_PASSWORD_LENGTH = 8
-
-  fun isCorrectName(name: String) = name.isNotEmpty()
-
-  fun isCorrectMail(mail: String): Boolean =
-      mail.isNotEmpty() && android.util.Patterns.EMAIL_ADDRESS.matcher(mail).matches()
-
-  fun isCorrectPassword(password: String): Boolean =
-      password.isNotEmpty() && password.length > MIN_PASSWORD_LENGTH
-
-  fun isCorrectCheckPassword(password: String, checkPassword: String) =
-      (password.isNotEmpty() && password == checkPassword)
-}
