+++ app/src/main/res/layout/fragment_signup.xml
@@ -146,6 +146,5 @@
         app:layout_constraintTop_toBottomOf="@+id/checkPasswordTextField"
         style="@style/ButtonStyle"
         />
-    />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
