+++ app/src/main/res/values/strings.xml
@@ -29,7 +29,7 @@
     <string name="title">Taskloans</string>
     <string name="sign_up_correctly">Signed up correctly</string>
     <string name="sign_in_correctly">Correct login</string>
-    <string name="sign_in">Entrar</string>
+    <string name="sign_in">Sign in</string>
     <string name="without_account">No account? <u>Create an account</u> </string>
 
     <string name="error_user_already_exists">Error, already exists a user with the same mail</string>
